Move distribution related methods to a concern
